Helper class for custom "native" memory images.
This is based on a sample of Harald Kuhr for memory mapped images. I.e. images that are
stored in "swap" files which get mapped into memory. The original sample is no longer
on GitHub, but you can find them here:

https://www.javatips.net/api/TwelveMonkeys-master/sandbox/sandbox-common/src/main/java/com/twelvemonkeys/image/MappedImageFactory.java

The version here is simplified and just uses native memory.

The version with memory mapped image files is very useful if you need to handle an
unknown amount of high-resolution images provided by customers, which may simply
not fit into main memory. Swapping them to disk allows to handle an nearly infinite
amount of images.