Delete invalid test cases for gray handling
When correctly handling color spaces, converting sRGB into
GRAY and then back again in sRGB is always going to be very
off.

GRAY is a linear color space with a linear gama curve.
sRGB has not a not linear gama curve (even if it behaves a little
bit like it has one). This means that there are effective less gray
values in sRGB as there are in GRAY. So you will always have conversion
artefacts.

When comparing image data you always have to be in the same color space.
When converting image data you should (if possible) always target a bigger
color space, so that you don't lose data. For RGB one very use full and big
color space is ProPhoto, which is complete linear. But ProPhoto requires
16 per channel as the color space is so big, that using only 8 bit per channel
will lead to massive steppings.